# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ultra Computer API
  version: 0.9.0
  description: |
    Full REST API for Ultra Computer — an AI orchestration platform with multi-agent
    swarm coordination, autonomous scheduling, voice, browser automation, marketplace,
    identity, messaging, and protocol interoperability (A2A, MCP, CLI, webhooks).
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
...
paths:
...
  /api/voice/transcribe:
    post:
      operationId: transcribeAudio
      summary: Transcribe audio to text (ASR) — JSON with base64 or raw binary body
      tags: [Voice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  description: Base64-encoded audio data
                format:
                  type: string
                  enum: [wav, opus, flac, mp3, pcm]
                  default: wav
                languageCode:
                  type: string
                enablePunctuation:
                  type: boolean
                  default: true
                enableTimestamps:
                  type: boolean
                maxAlternatives:
                  type: integer
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Transcription result
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcript:
                    type: string
                  alternatives:
                    type: array
                    items:
                      type: object
                  words:
                    type: array
                    items:
                      type: object
  /api/voice/synthesize:
    post:
      operationId: synthesizeSpeech
      summary: Synthesize text to speech (TTS)
      tags: [Voice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: Max 5000 characters
                voice:
                  type: string
                languageCode:
                  type: string
                sampleRate:
                  type: integer
                speed:
                  type: number
                pitch:
                  type: number
                volume:
                  type: number
                format:
                  type: string
                  enum: [wav, pcm, mp3]
                returnBase64:
                  type: boolean
                  description: If true, returns JSON with base64 audio; otherwise returns raw audio bytes
      responses:
        '200':
          description: Audio data (raw bytes or JSON with base64)
          content:
            audio/wav:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: object
                properties:
                  audio:
                    type: string
                    description: Base64-encoded audio
                  format:
                    type: string
                  sampleRate:
                    type: integer
                  durationMs:
                    type: integer
  /api/voice/pipeline:
    post:
      operationId: voicePipeline
      summary: Full voice pipeline — audio in → process → audio out
      tags: [Voice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  description: Base64-encoded input audio
                format:
                  type: string
                  default: wav
                languageCode:
                  type: string
                responseVoice:
                  type: string
                processingPrompt:
                  type: string
      responses:
        '200':
          description: Pipeline result
          content:
            application/json:
              schema:
                type: object
                properties:
                  inputTranscript:
                    type: string
                  processedText:
                    type: string
                  outputAudio:
                    type: string
                    description: Base64-encoded output audio
                  outputFormat:
                    type: string
                  totalLatencyMs:
                    type: integer
  /api/voice/voices:
    get:
      operationId: listVoices
      summary: List available TTS voices (cached 60s)
      tags: [Voice]
      parameters:
        - name: language
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Voice list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /api/voice/config:
    get:
      operationId: getVoiceConfig
      summary: Get current voice configuration
      tags: [Voice]
      responses:
        '200':
          description: Voice config
          content:
            application/json:
              schema:
                type: object
    patch:
      operationId: updateVoiceConfig
      summary: Update voice configuration
      tags: [Voice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                apiKey:
                  type: string
                nimEndpoint:
                  type: string
                defaultLanguage:
                  type: string
                defaultVoice:
                  type: string
                sampleRate:
                  type: integer
                enableStreaming:
                  type: boolean
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                type: object
...
  /api/voice/test:
    post:
      operationId: testVoice
      summary: Quick test — synthesize a test phrase and return base64 audio
      tags: [Voice]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                voice:
                  type: string
      responses:
        '200':
          description: Test result with base64 audio
          content:
            application/json:
              schema:
                type: object

  # ─── Marketplace ───────────────────────────────────────────────────────────
...